angel_oi_tracker/requirements.txt
# Core dependencies
mysql-connector-python==8.0.33
pandas==2.0.3
numpy==1.24.3
pytz==2023.3
apscheduler==3.10.1

# Angel One API
smartapi-python==1.3.0
pyotp==2.8.0

# Phase 4: AI Trade Intelligence
requests==2.31.0

# Phase 5: Dashboard Intelligence & API
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-multipart==0.0.6 